chore: extract keys_eq tests to dedicated test file
Moves the inline keys_eq and is_system_program unit tests from lib.rs
to lang/tests/keys_eq.rs. Keeps the test module focused and reduces
lib.rs noise.
